% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmajags-package.R
\docType{package}
\name{nmajags-package}
\alias{nmajags-package}
\alias{nmajags}
\title{nmajags: Brief overview of methods and general hints}
\description{
R package \bold{nmajags} is a user-friendly wrapper to conduct a
Bayesian network meta-analysis using JAGS.
}
\details{
R package \bold{nmajags} ...
}
\references{
Yu-Sung Su and Masanao Yajima, (2021):
R2jags: Using R to Run 'JAGS'.
https://CRAN.R-project.org/package=R2jags
}
\author{
Georgia Salanti \email{georgia.salanti@ispm.unibe.ch},
  Guido Schwarzer \email{sc@imbi.uni-freiburg.de}
}
